@@ -21,28 +21,27 @@
   },
   "require": {
     "aptoma/twig-markdown": "3.4.1",
-    "browscap/browscap-php": "7.4.0",
+    "browscap/browscap-php": "7.5.0",
     "cboden/ratchet": "0.4.4",
     "doctrine/annotations": "1.14.3",
     "doctrine/dbal": "3.8.3",
     "doctrine/orm": "2.19.6",
-    "egulias/email-validator": "4.0.2",
+    "egulias/email-validator": "4.0.3",
     "lcobucci/jwt": "5.3.0",
     "league/commonmark": "1.6.7",
-    "maennchen/zipstream-php": "3.1.0",
-    "onelogin/php-saml": "4.1.0",
+    "maennchen/zipstream-php": "3.1.1",
+    "onelogin/php-saml": "4.2.0",
     "php-ds/php-ds": "1.5.0",
     "ramsey/uuid": "4.7.6",
     "symfony/cache": "6.4.7",
     "symfony/config": "6.4.4",
-    "symfony/http-foundation": "6.3.2",
-    "symfony/polyfill-php80": "1.30.0",
+    "symfony/http-foundation": "6.4.14",
     "symfony/routing": "6.1.11",
     "textalk/websocket": "1.6.3",
-    "twig/twig": "3.10.3"
+    "twig/twig": "3.18.0"
   },
   "require-dev": {
-    "dealerdirect/phpcodesniffer-composer-installer": "0.7.2",
+    "dealerdirect/phpcodesniffer-composer-installer": "1.0.0",
     "php-mock/php-mock-phpunit": "2.9.0",
     "phpstan/phpstan": "1.10.57",
     "phpstan/phpstan-deprecation-rules": "1.1.4",
