@@ -20,13 +20,25 @@ use app\libraries\FileUtils;
 class RainbowCustomizationJSON extends AbstractModel {
     protected $core;
 
-    private $section;                   // Init in constructor
-    private $display_benchmark = [];
-    private $messages = [];
-    private $display = [];
-    private $benchmark_percent;         // Init in constructor
+    private object $section;                   // Init in constructor
+    /**
+     * @var string[]
+     */
+    private array $display_benchmark = [];
+    /**
+     * @var string[]
+     */
+    private array $messages = [];
+    /**
+     * @var string[]
+     */
+    private array $display = [];
+    private object $benchmark_percent;         // Init in constructor
     private object $final_cutoff;       // Init in constructor
-    private $gradeables = [];
+    /**
+     * @var object[]
+     */
+    private array $gradeables = [];
     /**
      * @var object[]
      */
@@ -75,7 +87,7 @@ class RainbowCustomizationJSON extends AbstractModel {
     /**
      * Get gradeable buckets array
      *
-     * @return array
+     * @return object[]
      */
     public function getGradeables() {
         return $this->gradeables;
@@ -84,7 +96,7 @@ class RainbowCustomizationJSON extends AbstractModel {
     /**
      * Get array of plagiarism
      *
-     * @return array<object>
+     * @return object[]
      */
     public function getPlagiarism(): array {
         return $this->plagiarism;
@@ -93,7 +105,7 @@ class RainbowCustomizationJSON extends AbstractModel {
     /**
      * Get array of manual grades
      *
-     * @return array<object>
+     * @return object[]
      */
     public function getManualGrades(): array {
         return $this->manual_grade;
@@ -111,7 +123,7 @@ class RainbowCustomizationJSON extends AbstractModel {
     /**
      * Gets an array of display benchmarks
      *
-     * @return array The display benchmarks
+     * @return string[] The display benchmarks
      */
     public function getDisplayBenchmarks() {
         return $this->display_benchmark;
@@ -137,7 +149,7 @@ class RainbowCustomizationJSON extends AbstractModel {
 
     /**
      * Gets an array of display
-     * @return array<string> The display
+     * @return string[] The display
      */
     public function getDisplay(): array {
         return $this->display;
@@ -152,12 +164,12 @@ class RainbowCustomizationJSON extends AbstractModel {
      * @param string $benchmark The benchmark to add
      * @throws BadArgumentException The passed in argument is not allowed
      */
-    public function addDisplayBenchmarks(string $benchmark) {
-        if (!in_array($benchmark, self::allowed_display_benchmarks)) {
+    public function addDisplayBenchmarks(string $benchmark): void {
+        if (!in_array($benchmark, self::allowed_display_benchmarks, true)) {
             throw new BadArgumentException('Passed in benchmark not found in the list of allowed benchmarks');
         }
 
-        if (!in_array($benchmark, $this->display_benchmark)) {
+        if (!in_array($benchmark, $this->display_benchmark, true)) {
             array_push($this->display_benchmark, $benchmark);
         }
     }
@@ -181,7 +193,7 @@ class RainbowCustomizationJSON extends AbstractModel {
      * @throws FileReadException Failure to read the contents of the file
      * @throws MalformedDataException Failure to decode the contents of the JSON string
      */
-    public function loadFromJsonFile() {
+    public function loadFromJsonFile(): void {
         // Get contents of file and decode
         $course_path = $this->core->getConfig()->getCoursePath();
         $course_path = FileUtils::joinPaths($course_path, 'rainbow_grades', 'gui_customization.json');
@@ -251,12 +263,12 @@ class RainbowCustomizationJSON extends AbstractModel {
      * @param string $display The item to add
      * @throws BadArgumentException The passed in argument is not allowed.
      */
-    public function addDisplay($display) {
-        if (!in_array($display, self::allowed_display)) {
+    public function addDisplay(string $display): void {
+        if (!in_array($display, self::allowed_display, true)) {
             throw new BadArgumentException('Passed in display not found in the list of allowed display items');
         }
 
-        if (!in_array($display, $this->display)) {
+        if (!in_array($display, $this->display, true)) {
             $this->display[] = $display;
         }
     }
@@ -266,14 +278,10 @@ class RainbowCustomizationJSON extends AbstractModel {
      *
      * @param string $sectionID The sectionID
      * @param string $label The label you would like to assign to the sectionID
-     * @throws BadArgumentException The passed in section label is empty
      */
-    public function addSection($sectionID, $label) {
-        if (empty($label)) {
-            throw new BadArgumentException('The section label may not be empty.');
-        }
-
-        $this->section->$sectionID = $label;
+    public function addSection(string $sectionID, string $label): void {
+        // If label is not set, use sectionID as default
+        $this->section->$sectionID = $label === '' ? $sectionID : $label;
     }
 
     /**
@@ -283,12 +291,12 @@ class RainbowCustomizationJSON extends AbstractModel {
      * @param float $percent The percent - this is the value for this json field
      * @throws BadArgumentException The passed in percent was empty
      */
-    public function addBenchmarkPercent($benchmark, $percent) {
-        if (empty($percent)) {
+    public function addBenchmarkPercent(string $benchmark, ?float $percent): void {
+        if (is_null($percent)) {
             throw new BadArgumentException('The benchmark percent may not be empty.');
         }
 
-        $this->benchmark_percent->$benchmark = (float) $percent;
+        $this->benchmark_percent->$benchmark = $percent;
     }
 
     /**
@@ -297,11 +305,12 @@ class RainbowCustomizationJSON extends AbstractModel {
      * @param string $cutoff The cutoff - this is the key for this json field
      * @param float $percent The percent - this is the value for this json field
      */
-    public function addFinalCutoff(string $cutoff, float $percent): void {
-        // To satisfy php-lint, add the pair to an array, then cast the array back to an object
-        $final_cutoff_array = (array) $this->final_cutoff;
-        $final_cutoff_array[$cutoff] = $percent;
-        $this->final_cutoff = (object) $final_cutoff_array;
+    public function addFinalCutoff(string $cutoff, ?float $percent): void {
+        if (is_null($percent)) {
+            throw new BadArgumentException('The final cutoff may not be empty.');
+        }
+
+        $this->final_cutoff->$cutoff = $percent;
     }
 
     /**
@@ -318,11 +327,11 @@ class RainbowCustomizationJSON extends AbstractModel {
      *
      * @param object $gradeable
      */
-    public function addGradeable(object $gradeable) {
+    public function addGradeable(object $gradeable): void {
         // Validation of this item will be better handled when schema validation is complete, until then just make
         // sure gradeable is not empty
         $emptyObject = (object) [];
-        if ($gradeable == $emptyObject) {
+        if ($gradeable === $emptyObject) {
             throw new BadArgumentException('Gradeable may not be empty.');
         }
 
@@ -332,7 +341,7 @@ class RainbowCustomizationJSON extends AbstractModel {
     /**
      * Get messages
      *
-     * @return array
+     * @return string[]
      */
     public function getMessages() {
         return $this->messages;
@@ -344,8 +353,8 @@ class RainbowCustomizationJSON extends AbstractModel {
      *
      * @param string $message
      */
-    public function addMessage(string $message) {
-        if (empty($message)) {
+    public function addMessage(string $message): void {
+        if ($message === '') {
             throw new BadArgumentException('You may not add an empty message.');
         }
 
@@ -434,13 +443,13 @@ class RainbowCustomizationJSON extends AbstractModel {
     /**
      * Save the contents in this objects properties to the customization.json for the current course
      */
-    public function saveToJsonFile() {
+    public function saveToJsonFile(): void {
         // Get path of where to save file
         $course_path = $this->core->getConfig()->getCoursePath();
         $course_path = FileUtils::joinPaths($course_path, 'rainbow_grades', 'gui_customization.json');
 
         // If display was empty then just add defaults
-        if (empty($this->display)) {
+        if (count($this->display) === 0) {
             $this->addDisplay('grade_summary');
             $this->addDisplay('grade_details');
         }
@@ -452,7 +461,7 @@ class RainbowCustomizationJSON extends AbstractModel {
         // @phpstan-ignore-next-line phpstan devs do not like object iteration
         foreach ($this as $key => $value) {
             // Dont include $core or $modified
-            if ($key != 'core' && $key != 'modified') {
+            if ($key !== 'core' && $key !== 'modified') {
                 $json->$key = $value;
             }
         }
